backend/requirements.txt
fastapi==0.107.0
uvicorn==0.22.0
requests==2.28.2
beautifulsoup4==4.11.1
pymongo==4.3.3
faiss-cpu==1.7.4
# Add torch explicitly (CPU wheel):
torch>=2.0
# Or don't specify exact version, at least set a minimum:
# torch>=2.0
sentence-transformers
huggingface-hub
transformers
kor==3.0.0
langchain==0.3.15
langchain_community==0.3.15
langchain==0.3.15
langchain-core==0.3.31
langchain-text-splitters==0.3.5
geopy==2.3.0
pyzill==1.0.0
openai==1.59.9
python-dotenv==1.0.0